Refactoring name files
