+++ src/works/antoniomaracil/components/vacation-form/components/vacation-form.js
@@ -1,340 +1,291 @@
 /* eslint-disable no-console */
 import { LitElement, html, css } from 'lit-element';
-import { nothing } from 'lit-html';
-import { getDate, formatDate, minMaxDate } from '../../../utils/functions';
+import { formatDate, getDate, checkDate } from './../../../utils/functions';
 
 class VacationForm extends LitElement {
   static get styles() {
-    return [
-      css`
-        .component-box{
-          margin: 1rem;
-          font-family: "Comic Sans MS", cursive, sans-serif;
-        }
-        .inp-controls{
-          display:flex;
-          align-items: center;
-        }
-        .inp-controls p{
-          margin-right: 1rem;
-        }
-        .inp-controls input+p{
-          margin-left: 1rem;
-        }
-        .inp-controls button{
-          margin-right: 1.5rem;
-          margin-left: 1.5rem;
-        }
-        .table-box{
-          border-top: solid 2px black;
-          border-bottom: solid 2px black;
-          margin-top: 1.5rem;
-          overflow-x: auto;
-        }
-        table {
-          border-collapse: collapse;
-          font-size: 0.8rem;
-          empty-cells: hide;
-          width: 100%;
-        }
-        tr:nth-child(even) {
-          background-color: #EEEEEE;
-        }
-        .inp-table th{
-          border-left: solid 2px black;
-          border-right: solid 2px black;
-          background-color: #CCCCCC;
-          font-size: 0.7rem;
-          text-align: left;
-        }
-        .inp-table td{
-          border-right: solid 2px black;
-          border-left: solid 2px black;
-        }
-        td{
-          height: 2rem;
-        }
-        span span{
-          cursor: pointer;
-        }
-        .table-cntr{
-          display: flex;
-          justify-content: space-between;
-        }
-        `
-    ];
+    return css`
+    .component-box{
+      margin: 1rem;
+      font-family: "Comic Sans MS", cursive, sans-serif;
+    }
+    .inp-controls{
+      display:flex;
+      align-items: center;
+    }
+    .inp-controls p{
+      margin-right: 1rem;
+    }
+    .inp-controls input+p{
+      margin-left: 1rem;
+    }
+    .inp-controls button{
+      margin-right: 1.5rem;
+      margin-left: 1.5rem;
+    }
+    .table-box{
+      border-top: solid 2px black;
+      border-bottom: solid 2px black;
+      margin-top: 1.5rem;
+    }
+    table {
+      border-collapse: collapse;
+      font-size: 0.8rem;
+      empty-cells: hide;
+      width: 100%;
+    }
+    tr:nth-child(even) {
+      background-color: #EEEEEE;
+    }
+    table th{
+      border-left: solid 2px black;
+      border-right: solid 2px black;
+      background-color: #CCCCCC;
+      font-size: 0.7rem;
+      text-align: left;
+    }
+    table td{
+      border-right: solid 2px black;
+      border-left: solid 2px black;
+    }
+    td{
+      height: 2rem;
+    }
+    span span{
+      cursor: pointer;
+    }
+    .table-cntr{
+      display: flex;
+      justify-content: space-between;
+    }
+
+    .stepper {
+      margin: 10px 0;
+    }
+
+    .stepper .step:hover {
+      background-color: #f1f1f1;
+    }
+
+    .step {
+      display: inline-block;
+      padding: 5px;
+      border: 1px solid #d8d7d7;
+      width: 20px;
+      height: auto;
+      text-align: center;
+      cursor: pointer;
+    }
+
+    .step.active {
+      background-color: #535353 !important;
+      color: white;
+    }
+
+    .step.left {
+      transform: rotate(180deg);
+    }
+
+    .stepper, .step {
+      user-select: none;
+    }
+    `;
   }
 
   static get properties() {
     return {
-      id: { type: Number },
-      pointer: { type: Number },
-      pagination: { type: Number },
-      vacation: { type: Object },
-      arrVacation: { type: Array },
-      arrTableView: { type: Array },
-      arrTr: { type: Array },
-      viewIsFull: { type: Boolean }
+      list: { type: Array },
+      nElements: { type: Number },
+      stepper: { type: Array, attribute: false },
+      table: { type: Array, attribute: false },
+      id: { type: Number, attribute: false },
+      index: { type: Number, attribute: false },
+      from: { type: Number, attribute: false },
+      to: { type: Number, attribute: false }
     };
   }
 
   constructor() {
     super();
+    this.list = [];
+    this.nElements = 3;
+    this.stepper = [];
+    this.from = 0;
+    this.to = this.nElements;
+    this.index = 0;
     this.id = 0;
-    this.pointer = 0;
-    this.pagination = 10;
-    this.vacation = {};
-    this.arrVacation = [];
-    this.arrTableView = [];
-    this.arrTr = new Array(10).fill({});
-    this.viewIsFull = false;
-    this.actualDate = new Date();
+    this.table = [];
   }
 
-  /**
-  * @desc Añade un nuevo registro al array de solicitudes
-  */
+  async firstUpdated() {
+    console.log(this.nElements);
+    this.table = new Array(this.nElements).fill({});
+    this.updateStepper();
+  }
+
+  async updateStepper() {
+    let nPages = Math.ceil(this.list.length / this.nElements);
+    if (nPages === 0) { nPages = 1; };
+    this.stepper = new Array(nPages).fill({});
+
+    await this.updateComplete;
+    this.index = this.stepper.length - 1;
+    this.setActiveStep(this.index);
+    this.showPage(this.index);
+  }
+
   add() {
-    // Actualizo la fecha para tener minutos actualizados
     this.actualDate = new Date();
     const start = getDate(this.shadowRoot.getElementById('start').value);
     const end = getDate(this.shadowRoot.getElementById('end').value);
-    // Comprobamos que las fechas tengan valor
-    const dateHasValue = start !== null && start !== '' && end !== null && end !== '';
-    if (dateHasValue) {
-      // Comprobamos que la fecha elegida sea mayor que la actual
-      if ((start.getTime() > this.actualDate.getTime())) {
-        // Comprobamos que la fecha de inicio sea menor que la de final
-        if ((start.getTime() < end.getTime())) {
-          this.vacation = {
-            id: this.id,
-            solicitud: formatDate(this.actualDate),
-            inicio: formatDate(start),
-            fin: formatDate(end),
-            estado: 'Pendiente de aprobación',
-            festado: formatDate(this.actualDate, true)
-          };
 
-          // Añadimos al array, incremento id
-          this.arrVacation.push(this.vacation);
-          this.id++;
+    if (checkDate(start, end)) {
+      this.vacation = {
+        id: this.id,
+        startDate: formatDate(start),
+        applicationDate: formatDate(this.actualDate),
+        endDate: formatDate(end),
+        status: 'Pendiente de aprobación',
+        statusDate: formatDate(this.actualDate, true)
+      };
 
-          // Actualizamos array de componente padre
-          this.sendData();
-
-          // Visualizar los datos
-          this.updateView();
-        }
-      }
+      this.list.push(this.vacation);
+      this.id++;
+      this.updateStepper();
     }
   }
 
-  /**
-  * @desc Actualiza el array de visualizacion
-  */
-  updateView() {
-    this.arrTableView = [];
-    const page = Math.trunc((this.arrVacation.length - 1) / this.pagination);
-    for (let i = 0; i < this.pagination; i++) {
-      if ((this.arrVacation.length - 1) < this.pagination) {
-        // La primera hoja se llena sin calcular la posicion del elemento
-        this.arrTableView[i] = this.arrVacation[i];
-      } else {
-        // Si la tabla esta llena volvemos true el flag
-        this.arrVacation.length % this.pagination === 0 ? this.viewIsFull : !this.viewIsFull;
-        // Calculo de la posicion del elemento
-        this.arrTableView[i] = this.arrVacation[(page * this.pagination) + i];
-      }
-    }
-    // Actualizo el puntero de la vista
-    this.pointer = Math.trunc(this.arrVacation.length / this.pagination);
-  }
-
-  /**
-  * @desc Elimina un registro del array y de la view
-  * @param String id del registro a eliminar
-  */
-  removeRow(e) {
-    // Buscamos la posicion de la id proprocionada por parametro
+  removeRow(id) {
     let index = 0;
-    for (let i = 0; i < this.arrVacation.length; i++) {
-      if (this.arrVacation[i].id === e) {
+    for (let i = 0; i < this.list.length; i++) {
+      if (this.list[i].id === id) {
         index = i;
         break;
       }
     }
+    const newList = [...this.list];
+    newList.splice(index, 1);
+    this.list = [...newList];
+    this.updateStepper();
+  }
 
-    // Sacamos del array
-    this.arrVacation.splice(index, 1);
-    // Mandamos el array actualizado al padre
-    this.sendData();
+  setActiveStep(index) {
+    this.shadowRoot.querySelectorAll('.step').forEach(row => {
+      if (row.id === `_${index}`) {
+        row.classList.add('active');
+      } else {
+        row.classList.remove('active');
+      }
+    });
+  }
 
-    // Comprobamos si la pagina a cambiado
-    const pageHasChanged =
-    this.pointer === Math.trunc((this.arrVacation.length - 1) / this.pagination);
+  showPage(index) {
+    this.index = index;
+    this.from = this.nElements * index;
+    this.to = this.from + this.nElements;
+    this.setActiveStep(index);
+  }
 
-    if (pageHasChanged) {
-      this.pointer = Math.trunc((this.arrVacation.length - 1) / this.pagination);
-    }
-
-    // Reiniciamos el array de visualizacion y rellenamos de nuevo
-    this.arrTableView = [];
-    for (let i = 0; i < this.pagination; i++) {
-      this.arrTableView[i] = this.arrVacation[(this.pointer * this.pagination) + i];
+  next() {
+    if (this.index < this.stepper.length - 1) {
+      this.showPage(this.index + 1);
     }
   }
 
-  /**
-  * @desc Mueve a izquierda y a derecha la view
-  * @param event
-  */
-  movePage(e) {
-    // Rectificamos el pointer si la pagina esta llena
-    if (this.viewIsFull) {
-      this.pointer--;
-      this.viewIsFull = false;
-    }
-
-    // Caso hacia la izquierda
-    if (e.target.id === 'left') {
-      this.pointer--;
-
-      // Comprobamos el pointer
-      if (this.pointer < 0) {
-        this.pointer = 0;
-      };
-
-      // Mostramos el array
-      this.arrTableView = [];
-      for (let i = 0; i < this.pagination; i++) {
-        this.arrTableView[i] = this.arrVacation[(this.pointer * this.pagination) + i];
-      }
-
-      // Caso hacia la derecha
-    } else if (e.target.id === 'right') {
-      this.pointer++;
-      // Comprobamos el pointer
-      if (this.pointer > Math.trunc((this.arrVacation.length - 1) / this.pagination)) {
-        this.pointer = Math.trunc((this.arrVacation.length - 1) / this.pagination);
-      }
-
-      // Mostramos el array
-      this.arrTableView = [];
-      for (let i = 0; i < this.pagination; i++) {
-        this.arrTableView[i] = this.arrVacation[(this.pointer * this.pagination) + i];
-      }
+  prev() {
+    if (this.index > 0) {
+      this.showPage(this.index - 1);
     }
   }
 
-  /**
-  * @desc Organiza el array segun fechas
-  * @param String col = propiedad a ordenar (f.solicitud, f.inicio o f.final)
-  * @param String order = 'asc' || 'desc'
-  */
-  dateOrg(col, order) {
-    this.arrVacation.sort((o1, o2) => {
+  orderList(column) {
+    const oldList = [...this.list];
+    const orderedList = this.list.sort((a, b) => {
       if (
-        getDate(o1[col], true).getTime() >
-        getDate(o2[col], true).getTime()
+        getDate(a[column], true).getTime() >
+        getDate(b[column], true).getTime()
       ) {
         return 1;
       } else if (
-        getDate(o1[col], true).getTime() <
-        getDate(o2[col], true).getTime()) {
+        getDate(a[column], true).getTime() <
+        getDate(b[column], true).getTime()) {
         return -1;
       }
       return 0;
     });
-
-    order === 'asc' ? this.arrVacation : this.arrVacation.reverse();
-
-    this.pointer = 0;
-    this.arrTableView = [];
-    for (let i = 0; i < 3; i++) {
-      this.arrTableView[i] = this.arrVacation[(this.pointer * 3) + i];
+    if (JSON.stringify(oldList) === JSON.stringify((orderedList))) {
+      orderedList.reverse();
     }
+    this.list = orderedList;
+    this.requestUpdate();
+    this.showPage(0);
   }
 
-  updated(changedProperties) {
-    changedProperties.forEach((oldValue, propName) => {
-      if (propName === 'arrVacation') {
-        this.updateView();
-      }
-    });
-  }
-
-  sendData() {
-    const event = new CustomEvent('update-array', {
-      detail: {
-        applications: this.arrVacation
-      }
-    });
-    this.dispatchEvent(event);
+  renderStepper() {
+    return html`
+      <div class="stepper">
+        <div class="step left" @click="${this.prev}">&#x25B7;</div>
+        ${this.stepper.map((x, i) => html`
+        <div id="${`_${i}`}" class="step" @click="${() => this.showPage(i)}">${i + 1}</div>
+        `)}
+        <div class="step" @click="${this.next}">&#x25B7;</div>
+      </div>
+    `;
   }
 
   render() {
     return html`
-        <div class="component-box">
-          <div class="inp-box">
-            <p>Solicitud de vacaciones:</p>
-            <div class="inp-controls">
-              <p>Fecha de inicio</p>
-              <input id="start" type="date" min="${minMaxDate(this.actualDate)}">
-              <p>Fecha de final</p>
-              <input id="end" type="date" max="${minMaxDate(this.actualDate, true)}">
-              <button @click="${this.add}">Agregar</button>
-            </div>
-          </div>
-        
-          <div class="table-cntr">
-            <button id="left" @click=${this.movePage}><</button> 
-            <button id="right" @click=${this.movePage}>></button>
-          </div>
-        
-          <div class="table-box">
-            <table class="inp-table">
-              <thead>
-                <tr id="head">
-                  <th>Fecha solicitud 
-                    <span>
-                      <span @click="${() => this.dateOrg('solicitud', 'asc')}">&#9652;</span>
-                      <span @click="${() => this.dateOrg('solicitud', 'desc')}">&#9662;</span>
-                    </span>
-                  </th>
-                  <th>Fecha inicio
-                  <span>
-                      <span @click="${() => this.dateOrg('inicio', 'asc')}">&#9652;</span>
-                      <span @click="${() => this.dateOrg('inicio', 'desc')}">&#9662;</span>
-                    </span>
-                  </th>
-                  <th>Fecha fin
-                  <span>
-                      <span @click="${() => this.dateOrg('fin', 'asc')}">&#9652;</span>
-                      <span @click="${() => this.dateOrg('fin', 'desc')}">&#9662;</span>
-                    </span>
-                  </th>
-                  <th>Estado solicitud</th>
-                  <th>Fecha estado</th>
-                  <th>Eliminar</th>
-                </tr>
-              </thead>
-              <tbody>
-              ${this.arrTr.map((value, i) => html`
-                <tr>
-                  <td>${this.arrTableView[i] ? this.arrTableView[i].solicitud : nothing}</td>
-                  <td>${this.arrTableView[i] ? this.arrTableView[i].inicio : nothing}</td>
-                  <td>${this.arrTableView[i] ? this.arrTableView[i].fin : nothing}</td>
-                  <td>${this.arrTableView[i] ? this.arrTableView[i].estado : nothing}</td>
-                  <td>${this.arrTableView[i] ? this.arrTableView[i].festado : nothing}</td>
-                  <td>${this.arrTableView[i] ? html`<button
-                      @click="${() => this.removeRow(this.arrTableView[i].id)}">Eliminar</button>` : nothing}</td>
-                </tr>             
-              `)}
-              </tbody>
-            </table>
+      <div class="container">
+      
+        <div class="inp-box">
+          <p>Solicitud de vacaciones:</p>
+          <div class="inp-controls">
+            <p>Fecha de inicio</p>
+            <input id="start" type="date">
+            <p>Fecha de final</p>
+            <input id="end" type="date">
+            <button @click="${this.add}">Agregar</button>
           </div>
         </div>
-      `;
+      
+        ${this.renderStepper()}
+        <div class="table-box">
+          <table>
+            <tr>
+              <th><button class="order" @click="${() => this.orderList('applicationDate')}">Fecha de solicitud
+                  <span>&#9662;</span></button></th>
+              <th><button class="order" @click="${() => this.orderList('startDate')}">Fecha de inicio
+                  <span>&#9662;</span></button></th>
+              <th><button class="order" @click="${() => this.orderList('endDate')}">Fecha de fin <span>&#9662;</span></button>
+              </th>
+              <th>Estado de solicitud</th>
+              <th>Fecha de solicitud</th>
+              <th>Eliminar</th>
+            </tr>
+            ${this.list.slice(this.from, this.to).map(item => html`
+            <tr>
+              <td>${item.applicationDate}</td>
+              <td>${item.startDate}</td>
+              <td>${item.endDate}</td>
+              <td>${item.status}</td>
+              <td>${item.statusDate}</td>
+              <td><button @click="${() => this.removeRow(item.id)}">Eliminar</button></td>
+            </tr>
+            `)}
+            ${this.table.map(item => html`
+            <tr>
+              <td></td>
+              <td></td>
+              <td></td>
+              <td></td>
+              <td></td>
+              <td></td>
+            </tr>
+            `)}
+          </table>
+        </div>
+      </div>
+    `;
   }
 }
